set up delivery and seller feedback

+++ Common/Workload/Metrics/Latency.cs
@@ -4,7 +4,7 @@
 	(
         int tid,
 		TransactionType type,
-		int period
+		double period
 	);
 }
